fix size bug

+++ src/consts/consts.hpp
@@ -5,6 +5,6 @@
 
 namespace nanofill::consts {
 
-using TradingDataCSVFormat = std::tuple<double, std::uint8_t, std::uint32_t, std::uint16_t, std::uint32_t, std::int8_t>;
+using TradingDataCSVFormat = std::tuple<double, std::uint8_t, std::uint32_t, std::uint32_t, std::uint32_t, std::int8_t>;
 
 }
