+++ src/events/event.hpp
@@ -50,14 +50,14 @@ struct Event {
     std::uint32_t time;
     std::uint32_t order_id;
     // Number of shares.
-    // 株の数。ネガティブなら、これは売り注文だ。
-    std::uint8_t size;
+    // 株の数。
+    std::uint32_t size;
     TradeDirection direction;
     EventType type;
 
     [[gnu::always_inline]]
-    std::int16_t get_size_with_direction() const noexcept {
-        return static_cast<std::int16_t>(size) * static_cast<std::int16_t>(direction);
+    std::int32_t get_size_with_direction() const noexcept {
+        return static_cast<std::int32_t>(size) * static_cast<std::int32_t>(direction);
     }
 };
